📦 refactor(test-folder): move test/ folder outside of src/
