+++ test/fixtures/installation-deleted-payload.json
