+++ test/unit/middleware-constructor-test.js
@@ -1,6 +1,6 @@
 const test = require("tap").test;
 
-const Middleware = require("../../middleware");
+const Middleware = require("../../src/middleware");
 
 test("options: none", (t) => {
   t.throws(Middleware);
